pwa-sync-starter/server/requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
python-dotenv==1.1.1
APScheduler==3.10.4
PyJWT==2.10.1
cryptography==45.0.7
requests==2.32.5
httpx==0.28.1
duckdb==1.1.3